requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.1.1
django-crispy-forms==2.3
django-jalali==7.0.0
djangorestframework==3.15.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
kavenegar==1.1.2
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
django-admin-interface==0.28.9